pkgdown/extra.html

Returns the URLs for all metadata files

Returns the URLs for all metadata files

Usage

get_metadata_url(
  databases = c("cellnexus_metadata.2.0.0.parquet", "census_cell_metadata.2.0.0.parquet")
)

Source

Mangiola et al.,2023

Arguments

databases

A character vector specifying the names of the metadata files. Download the specific metadata by defining the metadata version.

Value

A character vector of URLs to parquet files to download

References

Mangiola, S., M. Milton, N. Ranathunga, C. S. N. Li-Wai-Suen, A. Odainic, E. Yang, W. Hutchison et al. "A multi-organ map of the human immune system across age, sex and ethnicity." bioRxiv (2023): 2023-06. doi:10.1101/2023.06.08.542671.

Examples

get_metadata_url("cellnexus_metadata.2.0.0.parquet")
#> https://object-store.rc.nectar.org.au/v1/AUTH_06d6e008e3e642da99d806ba3ea629c5/cellNexus-metadata/cellnexus_metadata.2.0.0.parquet