pkgdown/extra.html

Clear the outdated metadata in the default cache directory.

Clear the outdated metadata in the default cache directory.

Usage

keep_updated_metadata(updated_data)

Source

Mangiola et al.,2023

Arguments

updated_data

A character vector of new metadata file name

Value

NULL, invisibly