Returns the default cache directory with a version number

Returns the default cache directory with a version number

get_default_cache_dir()

Source

Mangiola et al.,2023

Value

A length one character vector.

References

Mangiola, S., M. Milton, N. Ranathunga, C. S. N. Li-Wai-Suen, A. Odainic, E. Yang, W. Hutchison et al. "A multi-organ map of the human immune system across age, sex and ethnicity." bioRxiv (2023): 2023-06. doi:10.1101/2023.06.08.542671.

Examples

get_metadata(cache_directory = get_default_cache_dir())
#> # Source:   SQL [?? x 94]
#> # Database: DuckDB v1.3.2 [unknown@Linux 6.11.0-1018-azure:R 4.5.1/:memory:]
#>    cell_id                     dataset_id observation_joinid sample_id cell_type
#>    <chr>                       <chr>      <chr>              <chr>     <chr>    
#>  1 10X383_6:CCTCATGGTACACGCC_… 0ee5ae70-… *|Vt;4ATOI         e732d3bf… leukocyte
#>  2 10X383_6:GATTTCTTCCAACCAA_… 0ee5ae70-… #z5ft!kwbn         e732d3bf… leukocyte
#>  3 10X383_5:CTCCAACTCGCTCTAC_… 0ee5ae70-… kd(R>I7G#-         e732d3bf… leukocyte
#>  4 10X383_5:GATCATGCATGTGGTT_… 0ee5ae70-… mSylksfIn3         e732d3bf… leukocyte
#>  5 10X218_3:CAATCGATCCGATGTA_… 0ee5ae70-… XEl?wkHAcg         979206e1… leukocyte
#>  6 10X383_5:CTTCCTTCATGGCCCA_… 0ee5ae70-… T_4$MfQ!Ss         e732d3bf… leukocyte
#>  7 10X218_3:TATTGCTTCATCGCTC_… 0ee5ae70-… &yS%7ee<`?         979206e1… leukocyte
#>  8 10X218_3:CCATAAGGTGTCGATT_… 0ee5ae70-… 0*ArkI0ju;         979206e1… leukocyte
#>  9 10X218_3:ACTCTCGTCACCGCTT_… 0ee5ae70-… r<pyq>qS=W         979206e1… leukocyte
#> 10 10X383_6:CCCTTAGGTATTCTCT_… 0ee5ae70-… )tfjoT&}j{         e732d3bf… leukocyte
#> # ℹ more rows
#> # ℹ 89 more variables: cell_type_ontology_term_id <chr>, sample_ <chr>,
#> #   assay <chr>, assay_ontology_term_id <chr>, cell_count <chr>,
#> #   citation <chr>, collection_id <chr>, dataset_version_id <chr>,
#> #   default_embedding <chr>, development_stage <chr>,
#> #   development_stage_ontology_term_id <chr>, disease <chr>,
#> #   disease_ontology_term_id <chr>, donor_id <chr>, experiment___ <chr>, …